blogpost/requirements.txt
Django==1.6.6
djangorestframework==2.4.1
